requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
dacite==1.8.0
decorator==5.1.1
dnspython==2.3.0
docker==6.0.1
idna==3.4
msgpack==1.0.5
packaging==23.1
py==1.11.0
pymongo==4.3.3
PyYAML==6.0
requests==2.28.2
retry==0.9.2
urllib3==1.26.15
websocket-client==1.5.1
